Add validators
`Validator` instances can be added to an array of validators to process a
parser. Each `Validator` responds to `#validate` which accepts a `parser` and
returns an `Array` of `Validator::Error` structs. This `Array` is empty when the
parsed content is valid.

An generic (null) `Validator` is given as a base class. Additionally, an
`AlwaysInvalidValidator` (does what it says on the tin) is provided, mainly for
testing purposes. The outline of a `CsvFormatValidator` is also included here,
to be expanded later.