Fix `Parser` subclass registration
`@subclasses` was `nil` for subclasses of `Parser`, preventing them from being
inherited. This ensures that subclassing and subclass registration works
properly for nested subclasses by using a class variable, rather than a class
instance variable.

We also *prepend*, instead of append, classes to the array. This way the most
recently (which will tend to be the most locally) defined classes take
precedence when looking for parser class matches.