Add error streaming to `Validator`
Adds a single method interface for "ErrorStreams" (`#<<`). This is compatible
with `IO` and `Array`, which make for easy reuse of `STDOUT` and `Array` as easy
error stream classes. We will probably want to define a more complex
`ErrorStream` that accepts some kind of formattter for `Validation::Error`
etc...

Error output still needs to be handled for `RecordImporter`.

`Validator` subclasses now need to overwrite the private `#run_validation`
method to retain the default error streaming behavior.