Add error handling for unexpected RuntimeErrors during import
When running `RecordImporter#import` various errors are possible, through
`InputRecord`, or in identifying an import type, or otherwise during creation of
the object from the input record. When an error which is otherwise unhandled
reaches `#import` we send it to the `error_stream`, then re-raise.

Re-raising lets the error also reach the caller, which may have other handling,
or may want to abort further activity.