Add notifications before and after record creation
Adds an `info_stream` to `RecordImporter`, which defaults to `STDOUT`. This
stream can be used to process informational notices.

Initial notices are added to the `#import` call to let the user know that a
record is being processed, and when it has been created. The title and id are
included in the pre- and post-create notices, respectively.