Support validation of presence of `title` in records
This is a generic `title` presence validitor. It generates validation errors for
each record in the parser that does not have a title.

The validation is done at the `InputRecord` level, so the information given is
sparse (no line numbers), but the individual record metadata is reported with
each error to help identify the source of the issues.