Set depositor on new records. Story curationexperts/tenejo#137
* I renamed the `creator` instance variable to `depositor` in the
`HyraxRecordImporter` class.

* I changed the initializer in `HyraxRecordImporter` to use a keyword
argument called `attributes`, so that I could add another attribute for
the record.  In the future any new attributes for the record can be
passed in with that hash.

* If a depositor_id is passed in, the record's depositor will be set to
that user.

* I changed the `subject` importer in
`spec/integration/import_hyrax_csv.rb` to use a `HyraxRecordImporter`
instead of `RecordImporter`, since this is meant to be a test of a
hyrax-y import.