Pare down the update actor stack
When a record is imported a second time (i.e., when there is already a
record matching the deduplication field in the repository), only update
its metadata and collection membership. Use a stripped down actor stack
that only performs these actions.

Log a special message if we attempt an empty import.

Fix logging error where id number wasn't printing to logs on updates.